update text tests (strange)
