+++ apps/examples/e2e/tests/test-text.spec.ts
@@ -107,94 +107,6 @@ test.describe('text measurement', () => {
 		expect(formatLines(spans)).toEqual([['tes'], ['ting']])
 	})
 
-	test('for auto-font-sizing shapes, should do normal font size for text that does not have long words', async () => {
-		const shape = await page.evaluate(() => {
-			const id = 'shape:testShape' as TLShapeId
-			editor.createShapes([
-				{
-					id,
-					type: 'note',
-					x: 0,
-					y: 0,
-					props: {
-						text: 'this is just some regular text',
-						size: 'xl',
-					},
-				},
-			])
-
-			return editor.getShape(id) as TLNoteShape
-		})
-
-		expect(shape.props.fontSizeAdjustment).toEqual(32)
-	})
-
-	test('for auto-font-sizing shapes, should auto-size text that have slightly long words', async () => {
-		const shape = await page.evaluate(() => {
-			const id = 'shape:testShape' as TLShapeId
-			editor.createShapes([
-				{
-					id,
-					type: 'note',
-					x: 0,
-					y: 0,
-					props: {
-						text: 'Amsterdam',
-						size: 'xl',
-					},
-				},
-			])
-
-			return editor.getShape(id) as TLNoteShape
-		})
-
-		expect(shape.props.fontSizeAdjustment).toEqual(27)
-	})
-
-	test('for auto-font-sizing shapes, should auto-size text that have long words', async () => {
-		const shape = await page.evaluate(() => {
-			const id = 'shape:testShape' as TLShapeId
-			editor.createShapes([
-				{
-					id,
-					type: 'note',
-					x: 0,
-					y: 0,
-					props: {
-						text: 'this is a tentoonstelling',
-						size: 'xl',
-					},
-				},
-			])
-
-			return editor.getShape(id) as TLNoteShape
-		})
-
-		expect(shape.props.fontSizeAdjustment).toEqual(20)
-	})
-
-	test('for auto-font-sizing shapes, should wrap text that has words that are way too long', async () => {
-		const shape = await page.evaluate(() => {
-			const id = 'shape:testShape' as TLShapeId
-			editor.createShapes([
-				{
-					id,
-					type: 'note',
-					x: 0,
-					y: 0,
-					props: {
-						text: 'a very long dutch word like ziekenhuisinrichtingsmaatschappij',
-						size: 'xl',
-					},
-				},
-			])
-
-			return editor.getShape(id) as TLNoteShape
-		})
-
-		expect(shape.props.fontSizeAdjustment).toEqual(14)
-	})
-
 	test('should preserve whitespace at line breaks', async () => {
 		const spans = await page.evaluate<
 			{ text: string; box: BoxModel }[],
@@ -245,7 +157,7 @@ test.describe('text measurement', () => {
 			measureTextSpansOptions
 		)
 
-		expect(formatLines(spans)).toEqual([['  ', 'testing', ' ', 'testing']])
+		expect(formatLines(spans)).toEqual([['  ', 'testing', ' '], ['testing']])
 	})
 
 	test('should place starting whitespace on its own line if it has to', async () => {
@@ -329,5 +241,93 @@ test.describe('text measurement', () => {
 		)
 
 		expect(formatLines(spans)).toEqual([[' \n'], [' \n'], [' \n'], [' ']])
+	})
+
+	test('for auto-font-sizing shapes, should do normal font size for text that does not have long words', async () => {
+		const shape = await page.evaluate(() => {
+			const id = 'shape:testShape' as TLShapeId
+			editor.createShapes([
+				{
+					id,
+					type: 'note',
+					x: 0,
+					y: 0,
+					props: {
+						text: 'this is just some regular text',
+						size: 'xl',
+					},
+				},
+			])
+
+			return editor.getShape(id) as TLNoteShape
+		})
+
+		expect(shape.props.fontSizeAdjustment).toEqual(32)
+	})
+
+	test('for auto-font-sizing shapes, should auto-size text that have slightly long words', async () => {
+		const shape = await page.evaluate(() => {
+			const id = 'shape:testShape' as TLShapeId
+			editor.createShapes([
+				{
+					id,
+					type: 'note',
+					x: 0,
+					y: 0,
+					props: {
+						text: 'Amsterdam',
+						size: 'xl',
+					},
+				},
+			])
+
+			return editor.getShape(id) as TLNoteShape
+		})
+
+		expect(shape.props.fontSizeAdjustment).toEqual(28)
+	})
+
+	test('for auto-font-sizing shapes, should auto-size text that have long words', async () => {
+		const shape = await page.evaluate(() => {
+			const id = 'shape:testShape' as TLShapeId
+			editor.createShapes([
+				{
+					id,
+					type: 'note',
+					x: 0,
+					y: 0,
+					props: {
+						text: 'this is a tentoonstelling',
+						size: 'xl',
+					},
+				},
+			])
+
+			return editor.getShape(id) as TLNoteShape
+		})
+
+		expect(shape.props.fontSizeAdjustment).toEqual(21)
+	})
+
+	test('for auto-font-sizing shapes, should wrap text that has words that are way too long', async () => {
+		const shape = await page.evaluate(() => {
+			const id = 'shape:testShape' as TLShapeId
+			editor.createShapes([
+				{
+					id,
+					type: 'note',
+					x: 0,
+					y: 0,
+					props: {
+						text: 'a very long dutch word like ziekenhuisinrichtingsmaatschappij',
+						size: 'xl',
+					},
+				},
+			])
+
+			return editor.getShape(id) as TLNoteShape
+		})
+
+		expect(shape.props.fontSizeAdjustment).toEqual(14)
 	})
 })
