+++ packages/tldraw/src/lib/shapes/shared/TextHelpers.ts
@@ -4,8 +4,6 @@
  * Copyright (c) Federico Brigante <opensource@bfred.it> (bfred.it)
  */
 
-import { ANIMATION_MEDIUM_MS, Editor, TLShape } from '@tldraw/editor'
-
 // TODO: Most of this file can be moved into a DOM utils library.
 
 /** @internal */
@@ -291,43 +289,4 @@ function getCaretIndex(element: HTMLElement) {
 		position = preCaretRange.toString().length
 	}
 	return position
-}
-
-/** @internal */
-export function startEditingShapeWithLabel(editor: Editor, shape: TLShape, selectAll = false) {
-	// Finish this shape and start editing the next one
-	editor.select(shape)
-	editor.setEditingShape(shape)
-	editor.setCurrentTool('select.editing_shape', {
-		target: 'shape',
-		shape: shape,
-	})
-	if (selectAll) {
-		editor.emit('select-all-text', { shapeId: shape.id })
-	}
-	zoomToShapeIfOffscreen(editor)
-}
-
-const ZOOM_TO_SHAPE_PADDING = 16
-export function zoomToShapeIfOffscreen(editor: Editor) {
-	const selectionPageBounds = editor.getSelectionPageBounds()
-	const viewportPageBounds = editor.getViewportPageBounds()
-	if (selectionPageBounds && !viewportPageBounds.contains(selectionPageBounds)) {
-		const eb = selectionPageBounds
-			.clone()
-			// Expand the bounds by the padding
-			.expandBy(ZOOM_TO_SHAPE_PADDING / editor.getZoomLevel())
-			// then expand the bounds to include the viewport bounds
-			.expand(viewportPageBounds)
-
-		// then use the difference between the centers to calculate the offset
-		const nextBounds = viewportPageBounds.clone().translate({
-			x: (eb.center.x - viewportPageBounds.center.x) * 2,
-			y: (eb.center.y - viewportPageBounds.center.y) * 2,
-		})
-		editor.zoomToBounds(nextBounds, {
-			duration: ANIMATION_MEDIUM_MS,
-			inset: 0,
-		})
-	}
 }
